% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fizz_Buzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{Take in a number vector and return a vector of strings.}
\usage{
fizz_buzz(x)
}
\arguments{
\item{x}{A number}
}
\value{
A vector of strings of length x containing a number, Fizz, Buzz, or Fizz Buzz
}
\description{
Vector will contain 'Fizz' if index is divisible by 3, 'Buzz' if index is divisible
by 5, and 'Fizz Buzz' if index is divisible by both 3 and 5. Vector will contain
the index if it is not divisible by either.
}
\examples{
vector <- 1:17
fizz_buzz(vector)
}
