% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather data from the Flagstaff Pulliam Airport}
\format{
a data frame with 22967 observations of 5 columns.
\describe{
   \item{DATE}{The date of observation in POSIX date format.}
   \item{PRCP}{The amount of precipitation for the observed day in inches.}
   \item{SNOW}{The amount of snow for the observed day in inches.}
   \item{TempMax}{The max temperature in Fahrenheit for the observed day.}
   \item{TempMin}{The min temperature in Fahrenheit for the observed day.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
Weather data from the Flagstaff Pulliam Airport
}
\keyword{datasets}
